+++ stories/components/modal/modal.stories.ts
@@ -136,3 +136,88 @@ export const Sizes: StoryObj<ModalStoryArgs> = {
         showCloseButton: true
     }
 };
+
+/**
+ * **Control large content behavior**  
+ * When the content is too long, the modal can be customized with a max-height and an overflow. The CSS variables are :
+ * - `--wcs-modal-max-height` for controlling the max-height _(default: 80%)_
+ * - `--wcs-modal-overflow-y` for controlling the overflow-y _(default: auto)_
+ * - For `overflow-x`, adapt the modal size to your needs instead of adding a scrollbar
+ * _(see [Sizes story](.?path=/docs/components-modal--documentation#sizes))_
+ */
+export const OverflowAndMaxHeight: StoryObj = {
+    render: (args) => {
+        // FIXME
+        // @ts-ignore
+        const modalRef = createRef<HTMLWcsModalElement>();
+        return html`
+            <style>
+                .custom-variables {
+                    --wcs-modal-max-height: ${args["--wcs-modal-max-height"]};
+                    --wcs-modal-overflow-y: ${args["--wcs-modal-overflow-y"]};
+                }
+            </style>
+            <wcs-button @click="${_ => {
+                modalRef.value.show = true;
+            }}" id=${args.modalTriggerControlsId}>Show Modal ${args.size ? '(size: ' + args.size + ')' : null}
+            </wcs-button>
+            <wcs-modal ?show=${args.show}
+                       modal-trigger-controls-id=${args.modalTriggerControlsId}
+                       class="custom-variables"
+                       ${ref(modalRef)}
+                       ?without-backdrop=${args.withoutBackdrop}
+                       ?show-close-button=${args.showCloseButton}
+                       size=${ifDefined(args.size)}
+                       ?hide-actions=${args.hideActions}>
+                <div slot="header">Custom height and overflow modal</div>
+                <p><b>I have this in my CSS, edit me in the arguments !</b></p>
+                <ul>
+                    <li><code>--wcs-modal-max-height: 400px;</code></li>
+                    <li><code>--wcs-modal-overflow-y: auto;</code></li>
+                </ul>
+                <p>
+                    Lorem ipsum dolor sit amet, consectetur adipiscing elit. Sed non risus. Suspendisse lectus tortor,
+                    dignissim sit amet, adipiscing nec, ultricies sed, dolor. Cras elementum ultrices diam. Maecenas
+                    ligula massa, varius a, semper congue, euismod non, mi. Proin porttitor, orci nec nonummy molestie,
+                    enim est eleifend mi, non fermentum diam nisl sit amet erat. Duis semper. Duis arcu massa,
+                    scelerisque vitae, consequat in, pretium a, enim. Pellentesque congue. Ut in risus volutpat libero
+                    pharetra tempor. Cras vestibulum bibendum augue. Praesent egestas leo in pede. Praesent blandit odio
+                    eu enim. Pellentesque sed dui ut augue blandit sodales. Vestibulum ante ipsum primis in faucibus
+                    orci luctus et ultrices posuere cubilia Curae; Aliquam nibh. Mauris ac mauris sed pede pellentesque
+                    fermentum. Maecenas adipiscing ante non diam sodales hendrerit.
+                </p>
+                <p>
+                    Ut velit mauris, egestas sed, gravida nec, ornare ut, mi. Aenean ut orci vel massa suscipit
+                    pulvinar. Nulla sollicitudin. Fusce varius, ligula non tempus aliquam, nunc turpis ullamcorper nibh,
+                    in tempus sapien eros vitae ligula. Pellentesque rhoncus nunc et augue. Integer id felis. Curabitur
+                    aliquet pellentesque diam. Integer quis metus vitae elit lobortis egestas. Lorem ipsum dolor sit
+                    amet, consectetuer adipiscing elit. Morbi vel erat non mauris convallis vehicula. Nulla et sapien.
+                    Integer tortor tellus, aliquam faucibus, convallis id, congue eu, quam. Mauris ullamcorper felis
+                    vitae erat. Proin feugiat, augue non elementum posuere, metus purus iaculis lectus, et tristique
+                    ligula justo vitae magna.
+                </p>
+                <p>
+                    Curabitur eget fringilla ante. Interdum et malesuada fames ac ante ipsum primis in faucibus. Etiam
+                    efficitur mi vel luctus porta. Proin sollicitudin tempor nisi, nec dapibus ipsum blandit et. Aenean
+                    sit amet sem consectetur, malesuada enim at, condimentum ligula. Donec venenatis lobortis metus in
+                    pharetra. Curabitur dignissim lectus euismod aliquet ullamcorper. Suspendisse et nunc eget tortor
+                    scelerisque mattis. Curabitur a est finibus, maximus nisi eu, volutpat libero. Ut commodo massa
+                    diam, a tincidunt nibh luctus nec. Nam molestie volutpat nibh, eu finibus justo accumsan quis. Sed
+                    efficitur, turpis vel scelerisque lacinia, nibh est posuere lorem, efficitur dignissim erat turpis
+                    maximus neque. Donec consequat eros metus, vitae pharetra metus sollicitudin id. Praesent ut gravida
+                    erat. Donec nec ornare quam.
+                </p>
+                <div slot="actions">
+                    <wcs-button @click="${_ => { modalRef.value.show = false; }}" mode="stroked">Annuler</wcs-button>
+                    <wcs-button @click="${_ => {modalRef.value.show = false; }}">OK</wcs-button>
+                </div>
+            </wcs-modal>
+        `;
+    },
+    args: {
+        ...Default.args,
+        modalTriggerControlsId: "modal-trigger-controls-5",
+        "--wcs-modal-max-height": "400px",
+        "--wcs-modal-overflow-y": "auto"
+    }
+}